Updated navigation services, improved naming. Added a function call to the shared.
